+++ app/src/pages/project/TracesTable.tsx
@@ -120,9 +120,6 @@ const TableBody = <
                 <td
                   key={cell.id}
                   style={{
-                    // the cell still grows to fit, we just need some height declared
-                    // so that height: 100% works in children elements
-                    height: 1,
                     width: `calc(var(--col-${cell.column.id}-size) * 1px)`,
                     maxWidth: `calc(var(--col-${cell.column.id}-size) * 1px)`,
                     // prevent all wrapping, just show an ellipsis and let users expand if necessary